Investigate failure to read the whole input file
It turns out that the underlying warc library can silently fail to read most
of an input WARC file, at least if it's large and compressed. Outputs from a
simple read->write only contained the first c.3% of records.

On investigation, this seems to be due to a bug in the gzip processing code,
which is returning spurious EOFs. Although this may in principle be fixable,
the warc library appears to be abandoned. Work is ongoing on the warcutils
library, so it seems better to rebase warctika on top of warcutils.