Remove all traces of 'warc' library. All scripts now depend on hanzo's warc-utils.
The original 'warc' library, which warctika was original built
upon, was not a good choice. There were problems with reads from
large .warc.gz files being silently shortened, and many key
features were missing.

Everything has been rebased to the hanzo warc-utils library. As a
consequence, there is now no need for warctika to distribute a
full WARC handling library, so is now a collection of scripts which
do not need to be installed. The build architecture and library
files have therefore been removed.